+++Сart

+++ src/main/java/com.davyd.shop/service/UserService.java
@@ -2,9 +2,12 @@ package com.davyd.shop.service;
 
 import com.davyd.shop.dto.request.UserRequest;
 import com.davyd.shop.dto.response.UserResponse;
+import com.davyd.shop.entity.Cart;
+import com.davyd.shop.entity.Product;
 import com.davyd.shop.entity.User;
 import com.davyd.shop.exception.HasDependenciesException;
 import com.davyd.shop.exception.NoMatchesException;
+import com.davyd.shop.repository.CartRepository;
 import com.davyd.shop.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
@@ -20,18 +23,40 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private CartRepository cartRepository;
+
+    @Autowired
+    private ProductService productService;
+
     public void save(UserRequest request) {
         User user = new User();
         user.setUsername(request.getUsername());
         user.setPassword(request.getPassword());
+        Cart cart = new Cart();
+        cart = cartRepository.save(cart);
+        user.setCart(cart);
+
+        if (request.getFavoritesIds() != null) {
+            List<Product> collect = request.getFavoritesIds().stream()
+                    .map(productService::findOne)
+                    .collect(Collectors.toList());
+            user.setFavoriteProducts(collect);
+        }
         userRepository.save(user);
     }
+//    public void save(UserRequest request) {
+//        User user = new User();
+//        user.setUsername(request.getUsername());
+//        user.setPassword(request.getPassword());
+//        userRepository.save(user);
+//    }
     public List<UserResponse> findAll(String fieldName) {
         return userRepository.findAll(Sort.by(fieldName)).stream()
                 .map(UserResponse::new)
                 .collect(Collectors.toList());
     }
-    ///////////////////////////////////
+
     public User findOne(Long id) {
         return userRepository.findById(id).orElseThrow(() -> new NoMatchesException("User with id " + id + " not exists"));
     }
@@ -40,11 +65,11 @@ public class UserService {
     }
     public void delete(Long id) {
         User user = findOne(id);
-//        if (user.getComment().isEmpty()) {
+        if (user.getComments().isEmpty()) {
             userRepository.delete(user);
-//        } else {
-//            throw new HasDependenciesException("Cannot delete User with id " + id + " because it has comment");
-//        }
+        } else {
+            throw new HasDependenciesException("Cannot delete User with id " + id + " because it has comment");
+        }
     }
 
 
@@ -55,5 +80,24 @@ public class UserService {
         user.setUsername(request.getUsername());
         user.setPassword(request.getPassword());
         return user;
+    }
+
+    public void create(UserRequest request) {
+        User user = new User();
+        user.setPassword(request.getPassword());
+        user.setUsername(request.getUsername());
+
+        Cart cart = new Cart();
+        cart = cartRepository.save(cart);
+        user.setCart(cart);
+
+        if (request.getFavoritesIds() != null) {
+            List<Product> collect = request.getFavoritesIds().stream()
+                    .map(productService::findOne)
+                    .collect(Collectors.toList());
+            user.setFavoriteProducts(collect);
+        }
+
+        userRepository.save(user);
     }
 }
